+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Updated license year.
 - Updated Symfony to version 7.2.5.
 - `ChartExpression`: Internal update.
 - `ChartOptions`: Updated PHP documentation.